postCollectionViewをxibに
