+++ Nogistagram/Extension/NibInstantiatable.swift
@@ -14,6 +14,6 @@ extension UIView: NibInstantiatable {}
 extension NibInstantiatable where Self: UIView {
     static func instantiate(withOwner ownerOrNil: Any? = nil) -> Self {
         let nib = UINib(nibName: self.className, bundle: nil)
-        return nib.instantiate(withOwner: ownerOrNil, options: nil)[0] as! Self
+        return nib.instantiate(withOwner: ownerOrNil, options: nil).first as! Self
     }
 }
