+++ Nogistagram/ViewControllers/ExploreViewController.swift
@@ -12,15 +12,18 @@ import SwiftyJSON
 import ObjectMapper
 import Kingfisher
 
-class ExploreViewController: UIViewController, UICollectionViewDataSource, UICollectionViewDelegate {
+class ExploreViewController: UIViewController {
     
     var posts: [Post] = []
 
+    @IBOutlet weak var postCollectionWrapper: UIView!
     @IBOutlet weak var collectionView: UICollectionView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        loadPosts()
+        let postCollectionView = PostCollectionView.instantiate()
+        postCollectionView.frame = postCollectionWrapper.bounds
+        postCollectionWrapper.addSubview(postCollectionView) 
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -32,47 +35,4 @@ class ExploreViewController: UIViewController, UICollectionViewDataSource, UICol
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
-    func loadPosts() {
-        if let authDict = SessionHelper.authDict() {
-            Alamofire
-                .request("\(Constant.Api.root)/posts", method: .get, headers: authDict)
-                .validate(statusCode: 200..<300)
-                .responseJSON { response in
-                    switch response.result {
-                    case .success(let value):
-                        let postJsons = JSON(value)
-                        for (_, postJson) in postJsons {
-                            let post = Mapper<Post>().map(JSON: postJson.dictionaryObject!)!
-                            post.save()
-                        }
-                        self.posts = Post.all()
-                        self.collectionView.reloadData()
-                        self.collectionView.adaptBeautifulGrid(numberOfGridsPerRow: 3, gridLineSpace: 2.0)
-                    case .failure(let error):
-                        print(error)
-                    }
-                    
-            }
-        } else {
-            print("Error: authToken is nil")
-        }
-    }
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return posts.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "PostCollectionCell", for: indexPath) as! PostCollectionViewCell
-        let post = posts[indexPath.row]
-        cell.postImageView.kf.setImage(with: URL(string: post.imageUrl))
-        return cell
-    }
-    
-
 }
