+++ Nogistagram/Views/PostCollectionViewCell.swift
@@ -2,13 +2,19 @@
 //  PostCollectionViewCell.swift
 //  Nogistagram
 //
-//  Created by 村井謙太 on 2016/10/24.
+//  Created by 村井謙太 on 2016/10/27.
 //  Copyright © 2016年 村井謙太. All rights reserved.
 //
 
 import UIKit
 
 class PostCollectionViewCell: UICollectionViewCell {
-    
+
     @IBOutlet weak var postImageView: UIImageView!
+    
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
 }
