+++ Nogistagram/Views/PostCollectionViewCell.xib
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11201" systemVersion="15G31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11161"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="PostCollectionCell" id="gTV-IL-0wX" customClass="PostCollectionViewCell" customModule="Nogistagram" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="50" height="50"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
+                <rect key="frame" x="0.0" y="0.0" width="50" height="50"/>
+                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                <subviews>
+                    <imageView userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="OTD-Rz-qgp"/>
+                </subviews>
+            </view>
+            <constraints>
+                <constraint firstAttribute="trailing" secondItem="OTD-Rz-qgp" secondAttribute="trailing" id="1iM-u3-DB8"/>
+                <constraint firstItem="OTD-Rz-qgp" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" id="jon-CP-dxh"/>
+                <constraint firstItem="OTD-Rz-qgp" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" id="okf-Dk-D0s"/>
+                <constraint firstAttribute="bottom" secondItem="OTD-Rz-qgp" secondAttribute="bottom" id="r40-L4-a0H"/>
+            </constraints>
+            <connections>
+                <outlet property="postImageView" destination="OTD-Rz-qgp" id="NYw-Jg-lkw"/>
+            </connections>
+        </collectionViewCell>
+    </objects>
+</document>
